requirements.txt
asgiref==3.3.1
boto3==1.16.43
botocore==1.19.43
certifi==2020.12.5
chardet==4.0.0
Django==3.1.3
django-jsonfield==1.4.1
django-picklefield==3.0.1
django-post-office==3.5.3
django-storages==1.11.1
django-uuidfield-2==0.6.6
django-widget-tweaks==1.4.8
gunicorn==20.0.4
idna==2.10
jmespath==0.10.0
jsonfield==3.1.0
Pillow==8.0.1
psycopg2==2.8.6
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2020.4
requests==2.25.1
s3transfer==0.3.3
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
whitenoise==5.2.0
